Desktop/edconnekt/utilisateur-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
requests
python-dotenv
pydantic[email]